requirements.txt
streamlit
pymupdf
spacy
python-docx
matplotlib